Regroup conversion methods
