+++ src/lib.rs
@@ -3,6 +3,7 @@
 #![allow(clippy::let_and_return)]
 #![warn(missing_docs)]
 
+mod conversion;
 pub mod symbol;
 
 #[doc(hidden)]
@@ -203,57 +204,6 @@ impl Expr {
         }
     }
 
-    /// If this is a [`Normal`] expression, return that. Otherwise return None.
-    pub fn try_as_normal(&self) -> Option<&Normal> {
-        match self.kind() {
-            ExprKind::Normal(ref normal) => Some(normal),
-            ExprKind::Symbol(_)
-            | ExprKind::String(_)
-            | ExprKind::Integer(_)
-            | ExprKind::Real(_) => None,
-        }
-    }
-
-    /// If this is a [`True`] or [`False`] value, return that. Otherwise return None.
-    pub fn try_as_bool(&self) -> Option<bool> {
-        let s = self.try_as_symbol()?;
-        if s.eq("System`True") {
-            return Some(true);
-        }
-        if s.eq("System`False") {
-            return Some(false);
-        }
-        None
-    }
-
-    /// If this is a [`String`] expression, return that. Otherwise return None.
-    pub fn try_as_str(&self) -> Option<&str> {
-        match self.kind() {
-            ExprKind::String(ref string) => Some(string.as_str()),
-            _ => None,
-        }
-    }
-
-    /// If this is a [`Symbol`] expression, return that. Otherwise return None.
-    pub fn try_as_symbol(&self) -> Option<&Symbol> {
-        match self.kind() {
-            ExprKind::Symbol(ref symbol) => Some(symbol),
-            ExprKind::Normal(_)
-            | ExprKind::String(_)
-            | ExprKind::Integer(_)
-            | ExprKind::Real(_) => None,
-        }
-    }
-
-    /// If this is a [`Number`] expression, return that. Otherwise return None.
-    pub fn try_as_number(&self) -> Option<Number> {
-        match self.kind() {
-            ExprKind::Integer(int) => Some(Number::Integer(*int)),
-            ExprKind::Real(real) => Some(Number::Real(*real)),
-            ExprKind::Normal(_) | ExprKind::String(_) | ExprKind::Symbol(_) => None,
-        }
-    }
-
     /// Returns `true` if `self` is a `Normal` expr with the head `sym`.
     pub fn has_normal_head(&self, sym: &Symbol) -> bool {
         match *self.kind() {
@@ -266,7 +216,7 @@ impl Expr {
     // Common values
     //==================================
 
-    /// [`Null`](https://reference.wolfram.com/language/ref/Null.html)<sub>WL</sub>.
+    /// [`Null`](https://reference.wolfram.com/language/ref/Null.html) <sub>WL</sub>.
     pub fn null() -> Expr {
         Expr::symbol(unsafe { Symbol::unchecked_new("System`Null") })
     }
@@ -511,120 +461,6 @@ impl fmt::Display for Number {
                 let real: f64 = **real;
                 write!(f, "{:?}", real)
             },
-        }
-    }
-}
-
-//=======================================
-// Conversion trait impl's
-//=======================================
-
-impl From<Symbol> for Expr {
-    fn from(sym: Symbol) -> Expr {
-        Expr::symbol(sym)
-    }
-}
-
-impl From<&Symbol> for Expr {
-    fn from(sym: &Symbol) -> Expr {
-        Expr::symbol(sym)
-    }
-}
-
-impl From<Normal> for Expr {
-    fn from(normal: Normal) -> Expr {
-        Expr {
-            inner: Arc::new(ExprKind::Normal(normal)),
-        }
-    }
-}
-
-impl From<bool> for Expr {
-    fn from(value: bool) -> Expr {
-        match value {
-            true => Expr::symbol("System`True"),
-            false => Expr::symbol("System`False"),
-        }
-    }
-}
-
-macro_rules! string_like {
-    ($($t:ty),*) => {
-        $(
-            impl From<$t> for Expr {
-                fn from(s: $t) -> Expr {
-                    Expr::string(s)
-                }
-            }
-        )*
-    }
-}
-
-string_like!(&str, &String, String);
-
-//--------------------
-// Integer conversions
-//--------------------
-
-impl From<u8> for Expr {
-    fn from(int: u8) -> Expr {
-        Expr::from(i64::from(int))
-    }
-}
-
-impl From<i8> for Expr {
-    fn from(int: i8) -> Expr {
-        Expr::from(i64::from(int))
-    }
-}
-
-impl From<u16> for Expr {
-    fn from(int: u16) -> Expr {
-        Expr::from(i64::from(int))
-    }
-}
-
-impl From<i16> for Expr {
-    fn from(int: i16) -> Expr {
-        Expr::from(i64::from(int))
-    }
-}
-
-impl From<u32> for Expr {
-    fn from(int: u32) -> Expr {
-        Expr::from(i64::from(int))
-    }
-}
-
-impl From<i32> for Expr {
-    fn from(int: i32) -> Expr {
-        Expr::from(i64::from(int))
-    }
-}
-
-impl From<i64> for Expr {
-    fn from(int: i64) -> Expr {
-        Expr::number(Number::Integer(int))
-    }
-}
-
-// impl From<Normal> for ExprKind {
-//     fn from(normal: Normal) -> ExprKind {
-//         ExprKind::Normal(Box::new(normal))
-//     }
-// }
-
-// impl From<Symbol> for ExprKind {
-//     fn from(symbol: Symbol) -> ExprKind {
-//         ExprKind::Symbol(symbol)
-//     }
-// }
-
-impl From<Number> for ExprKind {
-    fn from(number: Number) -> ExprKind {
-        match number {
-            Number::Integer(int) => ExprKind::Integer(int),
-            Number::Real(real) => ExprKind::Real(real),
         }
     }
 }
